Move logger to its own file
